requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
